+++ packages/tx-parser/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.6](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/tx-parser@2.1.5...@radixdlt/tx-parser@2.1.6) (2021-07-22)
+
+**Note:** Version bump only for package @radixdlt/tx-parser
+
+
+
+
+
+## [2.1.5](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/tx-parser@2.1.4...@radixdlt/tx-parser@2.1.5) (2021-07-21)
+
+**Note:** Version bump only for package @radixdlt/tx-parser
+
+
+
+
+
+## [2.1.4](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/tx-parser@2.1.2...@radixdlt/tx-parser@2.1.4) (2021-07-21)
+
+**Note:** Version bump only for package @radixdlt/tx-parser
+
+
+
+
+
+## [2.1.3](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/tx-parser@2.1.2...@radixdlt/tx-parser@2.1.3) (2021-07-21)
+
+**Note:** Version bump only for package @radixdlt/tx-parser
+
+
+
+
+
 ## [2.1.2](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/tx-parser@2.1.1...@radixdlt/tx-parser@2.1.2) (2021-07-19)
 
 **Note:** Version bump only for package @radixdlt/tx-parser